feat(026-character-view): Combines character selection and character sheet views into single view. Updates Ancestry component to match naming convention.
