+++ data/characters/cheech.json
@@ -116,17 +116,17 @@
   ],
   "endurance": {
     "current": 10,
-    "max": 10
+    "max": 15
   },
   "hope": {
     "current": 12,
-    "max": 12
+    "max": 6
   },
   "defense": {
     "current": 13,
-    "max": 13
+    "max": 11
   },
-  "load": 4,
+  "load": 0,
   "shadow": 0,
   "injury": 0,
   "states": {
