+++ src/App.vue
@@ -3,7 +3,7 @@
     <div class="nav-menu" :class="{ open: menuOpen }">
       <button class="menu-toggle" @click="toggleMenu">☰</button>
       <nav v-if="menuOpen">
-        <router-link to="/character-selection" @click="closeMenu">CHARACTERS</router-link>
+        <router-link to="/characters" @click="closeMenu">CHARACTERS</router-link>
         <router-link to="/ancestries" @click="closeMenu">ANCESTRIES</router-link>
       </nav>
     </div>
