+++ src/stores/characterStore.js
@@ -11,9 +11,6 @@ export const useCharacterStore = defineStore('character', () => {
   const fetchCharacters = async () => {
     try {
       state.characters = await CharacterService.getAllCharacters();
-      if (state.characters.length > 0) {
-        state.selectedCharacter = state.characters[0]; // Default to first character
-      }
     } catch (error) {
       console.error("Failed to fetch characters:", error);
     }
